refactor: Separate UI Logic from domain
