+++ src/main/java/calculator/Application.java
@@ -1,9 +1,10 @@
 package calculator;
 
 import calculator.domain.Parser;
+import calculator.view.ResultView;
 import java.util.List;
 import calculator.domain.Calculator;
-import calculator.domain.Input;
+import calculator.view.InputView;
 
 
 public class Application {
@@ -11,10 +12,11 @@ public class Application {
         System.out.println("연산 가능한 수식을 입력해주세요. (숫자, 연산자, 공백 문자만 허용됩니다.)");
         while (true) {
             try {
-                String userInput = Input.getInput();
+                String userInput = InputView.getInput();
                 List<String> parsedInput = Parser.parseInput(userInput);
                 Calculator calculator = Calculator.getInstance();
-                calculator.calculate(parsedInput);
+                int result = calculator.calculate(parsedInput);
+                ResultView.printResult(result);
                 break;
             } catch(IllegalArgumentException | ArithmeticException e) {
                 System.out.println("[ERROR] 연산 가능한 수식을 입력해주세요. (숫자, 연산자, 공백 문자만 허용됩니다.)");
