+++ src/main/java/calculator/view/InputView.java
@@ -1,8 +1,9 @@
-package calculator.domain;
+package calculator.view;
 
+import calculator.domain.Validator;
 import java.util.Scanner;
 
-public class Input {
+public class InputView {
 
     public static String getInput() {
         String userInput;
