+++ src/main/java/calculator/view/ResultView.java
@@ -1,0 +1,7 @@
+package calculator.view;
+
+public class ResultView {
+    public static void printResult(int result) {
+        System.out.println(result);
+    }
+}
